feat: support custom instance names for consistent hashing
When using the consistent hashing routing provided by kitex,
if the instance addresses obtained from service discovery change,
it causes rebalancing, which means recalculating
the request routing to instances.
However, in some cases, although the IP addresses change,
the instances themselves remain the same (for example,
update stateful containers in Kubernetes with local disks
allocated and some cache data stored).
Therefore, we want to use consistent hashing to route requests to the same containers.

In this PR, we add a GetName function to the Consistent Loadbalancer
creation options. At the same time,
when calculating the consistent hash key for downstream instances,
the function attempts to obtain the instance name associated with
the IP addresses, rather than using the IP addresses directly.
In this way, consistent hashing can still be achieved
even if the IP addresses change.

Signed-off-by: Yorling <shallowcloud@yeah.net>