Created jsonpb_codec.go, which encapsulates the generic. Notable differences include using the PbDescriptorProvider as the provider
Additionally modified the core Message interface to have a SetData()
field

In the Marshal method in jsonpb_codec.go, I parsed the json and created
a dynamic.Message, then replaced the contents of Data in the message
with the dynamic.Message. I then modified protobufs.go's Marshal method
to check for the dynamic.Message using type assertions, then marshalled
the dynamic.Message and wrote it to the output byte buffer.

Currently this results in an error
[Error] KITEX: processing request error, remoteService=, remoteAddr=127.0.0.1:53434, error=Bad version in ReadMessageBegin
Which I am still trying to resolve

Tests can be located in pkg/generic/generic_test.go, and
/generic/jsonpb_test

Modifications for unmarshal methods have not been implemented since I'm
trying to resolve the encoding issues